trx兌換機器人

+++ src/main/java/com/yt/app/api/v1/entity/Channel.java
@@ -36,7 +36,7 @@ public class Channel extends YtBaseEntity<Channel> {
 	Double todaycost;
 	Double todaycount;
 	Double count;
-	Integer collection;
+	Double collection;
 	Double exchangeonecost;
 	Double exchangedownpoint;
 	Double downpoint;
