+++ src/main/java/com/yt/app/api/v1/entity/Merchantqrcodeaisle.java
@@ -26,6 +26,6 @@ public class Merchantqrcodeaisle extends YtBaseEntity<Merchantqrcodeaisle> {
 	Long tenant_id;
 	Long merchantid;
 	Long qrcodeaisleid;
-	Integer collection;
+	Double collection;
 	Integer version;
 }
